requirements.txt
mysql_connector
matplotlib
tk_tools